#10_punktestand
Added logic for querying score of a user in the endpoint service, which is called by the new points component.
The service delivers the score in a numbers array, but there is something to change because only the score of answers that match the tags of the challenge must be tracked.
For test purposes three .json files have been added to simulate a multi-page query from the StackExchange-API.
Chart.js and ng2-charts have been installed to use it furthermore.
Model class 'answer' has been added that encapsulates fields for the above mentioned changes that have to be applied later on.
New route to the score's page introduced (may be deleted later, it is likely that it is not necessary later on)